feat(events): refactored interfaces in separate file
