+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './request';
 export * from './shared';
 export * from './v5-shared';
 export * from './websockets';
+export * from './websocket.events';
